Halve memory usage of saved feedrate
Truncate the saved feedrate to an uint16_t.
This is more than sufficient for recovery.

+++ Firmware/planner.h
@@ -118,7 +118,7 @@ typedef struct {
 
   // Save/recovery state data
   float gcode_target[NUM_AXIS];     // Target (abs mm) of the original Gcode instruction
-  float gcode_feedrate;             // Original feedrate
+  uint16_t gcode_feedrate;          // Default and/or move feedrate
   uint16_t sdlen;                   // Length of the Gcode instruction
 } block_t;
